Conflicts:
	.classpath

+++ src/cube/Cube.java
@@ -1,26 +1,18 @@
 package cube;
 
-import utilities.Point;
-
-
-//import Xbee library
-import com.rapplogic.xbee.api.XBee;
-import com.rapplogic.xbee.api.XBeeAddress64;
-import com.rapplogic.xbee.api.XBeeException;
-import com.rapplogic.xbee.api.zigbee.ZNetTxRequest;
-import com.rapplogic.xbee.util.ByteUtils;
-
-import processing.serial.*;//import Xbee library */
+//import processing.serial.*;//import Xbee library */
 
 
 public class Cube {	
+
 	/**
 	 * Attribute of cube
 	 */
 	private short angle;
 	private boolean tap;
+
 	private short id;
-	
+
 	/**
 	 * Method to set the RGB color of the LEDS
 	 * @parameters : chart of 
@@ -37,7 +29,7 @@ public class Cube {
 		
 	}
 	
-	public short getAngle() {
+	public float getAngle() {
 		return angle;
 	}
 	
@@ -47,5 +39,6 @@ public class Cube {
 	
 	public short getID() {
 		return id;
-	}
+
+}
 }
